+++ dusty/commands/test.py
@@ -12,7 +12,7 @@ from ..systems.docker.compose import write_composefile, compose_up
 from ..systems.rsync import sync_repos_by_specs
 from ..systems.virtualbox import initialize_docker_vm
 from ..log import log_to_client
-from ..command_file import make_test_command_files, remove_test_command_files, dusty_command_file_name
+from ..command_file import make_test_command_files, dusty_command_file_name
 
 def test_info_for_app_or_lib(app_or_lib_name):
     expanded_specs = get_expanded_libs_specs()
@@ -34,27 +34,23 @@ def run_app_or_lib_tests(app_or_lib_name, suite_name, test_arguments, force_recr
     client = get_docker_client()
     expanded_specs = get_expanded_libs_specs()
     spec = expanded_specs.get_app_or_lib(app_or_lib_name)
-    sync_repos_by_specs([spec])
     test_command = _construct_test_command(spec, suite_name, test_arguments)
+    make_test_command_files(expanded_specs)
+    sync_repos_by_specs([spec])
     ensure_test_image(client, app_or_lib_name, expanded_specs, force_recreate=force_recreate)
     _run_tests_with_image(client, expanded_specs, app_or_lib_name, test_command)
 
 def _construct_test_command(spec, suite_name, test_arguments):
-    suite_command = None
+    suite_spec = None
     for suite_dict in spec['test']['suites']:
         if suite_dict['name'] == suite_name:
-            suite_command = '; '.join(suite_dict['command'])
-            suite_default_args = suite_dict['default_args']
+            suite_spec = suite_dict
             break
-    if suite_command is None:
+    if suite_spec is None:
         raise RuntimeError('{} is not a valid suite name'.format(suite_name))
     if not test_arguments:
-        test_arguments = suite_default_args.split(' ')
-    cd_command = 'cd {}'.format(container_code_path(spec))
-    sub_command = "{} {}".format(suite_command, ' '.join(test_arguments))
-    test_command = 'sh -c "{}; {}"'.format(cd_command, sub_command.strip())
-    log_to_client('Command to run in test is {}'.format(test_command))
-    return test_command
+        test_arguments = suite_spec['default_args'].split(' ')
+    return 'sh {}/{} {}'.format(constants.CONTAINER_COMMAND_FILES_DIR, dusty_command_file_name(spec.name, test_name=suite_name), ' '.join(test_arguments))
 
 def _test_composefile_path(service_name):
     return os.path.expanduser('~/.dusty-testing/test_{}.yml'.format(service_name))
@@ -96,7 +92,7 @@ def _app_or_lib_compose_up(testing_spec, app_or_lib_name, app_or_lib_volumes, te
 def _run_tests_with_image(client, expanded_specs, app_or_lib_name, test_command):
     testing_spec = expanded_specs.get_app_or_lib(app_or_lib_name)['test']
 
-    volumes = get_volume_mounts(app_or_lib_name, expanded_specs)
+    volumes = get_volume_mounts(app_or_lib_name, expanded_specs, test=True)
     previous_container_names = _services_compose_up(expanded_specs, app_or_lib_name, testing_spec)
     previous_container_name = previous_container_names[-1] if previous_container_names else None
     test_container_name = _app_or_lib_compose_up(testing_spec, app_or_lib_name,