+++ dusty/path.py
@@ -10,3 +10,7 @@ def parent_dir(path):
 
 def vm_cp_path(app_or_service_name):
     return os.path.join(constants.VM_CP_DIR, app_or_service_name)
+
+def vm_command_files_path(app_or_service_name):
+    return os.path.join(constants.VM_COMMAND_FILES_DIR, app_or_service_name)
+