+++ dusty/systems/docker/testing_image.py
@@ -5,6 +5,8 @@ import docker
 from ...compiler.compose import container_code_path, get_volume_mounts
 from ...log import log_to_client
 from ...command_file import dusty_command_file_name, lib_install_commands_for_app_or_lib
+from .common import spec_for_service
+from ... import constants
 
 def _ensure_testing_spec_base_image(docker_client, testing_spec):
     log_to_client('Getting the base image for the new image')
@@ -66,36 +68,21 @@ def _make_installed_requirements_image(docker_client, base_image_tag, command, i
     new_image = docker_client.commit(container=container['Id'])
     docker_client.tag(image=new_image['Id'], repository=image_name, force=True)
 
-def _spec_for_service(app_or_lib_name, expanded_specs):
-    if app_or_lib_name in expanded_specs['apps']:
-        return expanded_specs['apps'][app_or_lib_name]
-    return expanded_specs['libs'][app_or_lib_name]
-
 def _testing_spec(app_or_lib_name, expanded_specs):
-    return _spec_for_service(app_or_lib_name, expanded_specs)['test']
+    return spec_for_service(app_or_lib_name, expanded_specs)['test']
 
 def test_image_name(app_or_lib_name):
     return "dusty/test_{}".format(app_or_lib_name)
 
-def _get_testing_spec_command(testing_spec):
-    command = ""
-    for line in testing_spec['once']:
-        command += "{};".format(line.replace('"', '\\"'))
-    return command
-
-def _get_test_image_setup_command(app_or_lib_name, expanded_specs):
-    testing_spec = _testing_spec(app_or_lib_name, expanded_specs)
-    commands = lib_install_commands_for_app_or_lib(app_or_lib_name, expanded_specs)
-    commands += ['cd {}'.format(container_code_path(_spec_for_service(app_or_lib_name, expanded_specs)))]
-    commands += [_get_testing_spec_command(testing_spec)]
-    return "sh -c \"{}\"".format('; '.join(commands))
+def _get_test_image_setup_command(app_or_lib_name, app_or_lib_spec):
+    return 'sh {}/{}'.format(constants.CONTAINER_COMMAND_FILES_DIR, dusty_command_file_name(app_or_lib_name))
 
 def _make_installed_testing_image(docker_client, app_or_lib_name, expanded_specs):
     image_name = test_image_name(app_or_lib_name)
     testing_spec = _testing_spec(app_or_lib_name, expanded_specs)
     base_image_tag = _ensure_testing_spec_base_image(docker_client, testing_spec)
-    image_setup_command = _get_test_image_setup_command(app_or_lib_name, expanded_specs)
-    volumes = get_volume_mounts(app_or_lib_name, expanded_specs)
+    image_setup_command = _get_test_image_setup_command(app_or_lib_name, spec_for_service(app_or_lib_name, expanded_specs))
+    volumes = get_volume_mounts(app_or_lib_name, expanded_specs, test=True)
     _make_installed_requirements_image(docker_client, base_image_tag, image_setup_command, image_name, volumes)
 
 def ensure_test_image(docker_client, app_or_lib_name, expanded_specs, force_recreate=False):